Fix sub-pixel rendering
- Switch to storing position and velocity in integer units.
- Limit updating to once a tick.
- Render tiles and sprites in pixel-space instead of tile-space.
- Fix stat box player name sorting so it's not case sensitive.

+++ src/PlayerManager.h
@@ -33,7 +33,7 @@ struct PlayerManager {
   void OnLargePositionPacket(u8* pkt, size_t size);
   void OnSmallPositionPacket(u8* pkt, size_t size);
 
-  void OnPositionPacket(Player& player, const Vector2f& position);
+  void OnPositionPacket(Player& player, const Vector2i& position);
 };
 
 }  // namespace null
